requirements.txt
requests~=2.31.0
beautifulsoup4~=4.12.2
telepot~=12.7
PyYAML~=6.0